scraper/requirements.txt
flask==1.0.2
beautifulsoup4==4.6.3
requests==2.21.0
urllib==1.24.1
lask-restplus==0.13.0
PyYAML